requirements.txt
arrow==1.2.2
asgiref==3.5.0
asttokens==2.0.5
atomicwrites==1.4.0
attrs==21.4.0
backcall==0.2.0
beautifulsoup4==4.11.1
binaryornot==0.4.4
cachetools==4.2.4
certifi==2021.10.8
chardet==4.0.0
charset-normalizer==2.0.12
click==8.1.2
colorama==0.4.4
cookiecutter==1.7.3
debugpy==1.6.0
decorator==4.4.2
Django==4.0.2
django-browser-reload==1.3.0
django-crispy-forms==1.14.0
django-tailwind==3.1.1
django-unicorn==0.44.1
djangorestframework==3.13.1
entrypoints==0.4
executing==0.8.3
fonttools==4.31.2
gunicorn==20.1.0
idna==3.3
iniconfig==1.1.1
ipykernel==6.10.0
ipython==8.2.0
jedi==0.18.1
Jinja2==3.1.1
jinja2-time==0.2.0
jupyter-client==7.2.0
jupyter-core==4.9.2
kiwisolver==1.4.0
MarkupSafe==2.1.1
matplotlib-inline==0.1.3
MouseInfo==0.1.3
nest-asyncio==1.5.4
npm==0.1.1
numpy==1.22.3
optional-django==0.1.0
orjson==3.6.7
packaging==21.3
pandas==1.4.1
parso==0.8.3
pickleshare==0.7.5
Pillow==9.0.1
plotly==5.7.0
pluggy==1.0.0
poyo==0.5.0
prompt-toolkit==3.0.28
psutil==5.9.0
psycopg2==2.9.3
pure-eval==0.2.2
py==1.11.0
PyAutoGUI==0.9.53
PyGetWindow==0.0.9
Pygments==2.11.2
PyMsgBox==1.0.9
pyparsing==3.0.7
pyperclip==1.8.2
PyRect==0.2.0
PyScreeze==0.1.28
pytest==7.0.0
python-dateutil==2.8.2
python-slugify==6.1.1
pytweening==1.0.4
pytz==2021.3
# pywin32==303
pyzmq==22.3.0
requests==2.27.1
shortuuid==1.0.8
six==1.16.0
soupsieve==2.3.2
sqlparse==0.4.2
stack-data==0.2.0
tenacity==8.0.1
text-unidecode==1.3
tomli==2.0.0
tornado==6.1
traitlets==5.1.1
tzdata==2021.5
urllib3==1.26.8
wcwidth==0.2.5
whitenoise==6.0.0
